fix: correçao na leitura de dados de educaçao, saude e transporte

+++ src/main/java/com/databrick/entity/EducationSystem.java
@@ -4,19 +4,19 @@ public class EducationSystem {
     private String schoolName;
     private String neighborhood;
     private String districtName;
-    private Integer districtCode;
+    private Long districtCode;
 
     public EducationSystem(String schoolName, String neighborhood, String districtName, String districtCode) {
         this.schoolName = schoolName;
         this.neighborhood = neighborhood;
         this.districtName = districtName;
-        this.districtCode = parseInteger(districtCode);
+        this.districtCode = parseLong(districtCode);
     }
 
-    private Integer parseInteger(String value) {
+    private Long parseLong(String value) {
         if (value == null || value.isBlank()) return null;
         try {
-            return Integer.valueOf(value.replace(",", "").replace(".", ""));
+            return Long.valueOf(value.replace(",", "").replace(".", ""));
         } catch (NumberFormatException e) {
             return null;
         }
@@ -46,11 +46,11 @@ public class EducationSystem {
         this.districtName = districtName;
     }
 
-    public Integer getDistrictCode() {
+    public Long getDistrictCode() {
         return districtCode;
     }
 
-    public void setDistrictCode(Integer districtCode) {
+    public void setDistrictCode(Long districtCode) {
         this.districtCode = districtCode;
     }
 }
