+++ src/main/java/com/databrick/entity/HealthCare.java
@@ -1,34 +1,34 @@
 package com.databrick.entity;
 
 public class HealthCare {
-    private Integer districtCode;
+    private Long districtCode;
     private String districtName;
     private String unitName;
     private String neighborhood;
     private String unitType;
 
     public HealthCare(String districtCode, String districtName, String unitName, String unitType, String neighborhood) {
-        this.districtCode = parseInteger(districtCode);
+        this.districtCode = parseLong(districtCode);
         this.districtName = districtName;
         this.unitName = unitName;
         this.unitType = unitType;
         this.neighborhood = neighborhood;
     }
 
-    private Integer parseInteger(String value) {
+    private Long parseLong(String value) {
         if (value == null || value.isBlank()) return null;
         try {
-            return Integer.valueOf(value.replace(",", "").replace(".", ""));
+            return Long.valueOf(value.replace(",", "").replace(".", ""));
         } catch (NumberFormatException e) {
             return null;
         }
     }
 
-    public Integer getDistrictCode() {
+    public Long getDistrictCode() {
         return districtCode;
     }
 
-    public void setDistrictCode(Integer districtCode) {
+    public void setDistrictCode(Long districtCode) {
         this.districtCode = districtCode;
     }
